chore: track changes in ibc-rs PR 1115
